tree: refactor contextuallyTyped out of editable-tree
