+++ packages/dds/tree/src/feature-libraries/editable-tree/index.ts
@@ -5,7 +5,6 @@
 
 export {
 	typeSymbol,
-	typeNameSymbol,
 	EditableTree,
 	EditableField,
 	EditableTreeOrPrimitive,
@@ -14,7 +13,6 @@ export {
 	proxyTargetSymbol,
 	UnwrappedEditableTree,
 	UnwrappedEditableField,
-	valueSymbol,
 	indexSymbol,
 	getField,
 	createField,
@@ -24,14 +22,4 @@ export {
 
 export { EditableTreeContext, getEditableTreeContext } from "./editableTreeContext";
 
-export {
-	PrimitiveValue,
-	isPrimitiveValue,
-	isPrimitive,
-	getPrimaryField,
-	ContextuallyTypedNodeDataObject,
-	ContextuallyTypedNodeData,
-	MarkedArrayLike,
-	isWritableArrayLike,
-	isContextuallyTypedNodeDataObject,
-} from "./utilities";
+export { isPrimitive } from "./utilities";
